ResNet18-CIFAR10-Image-Classifier/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
torch==2.0.1
torchvision==0.15.2
tensorflow==2.15.0
Pillow==10.0.0
numpy==1.26.2
python-multipart>=0.0.9
gradio==4.44.1
huggingface_hub==0.24.0